requirements.txt
mysql-connector-python==8.0.13
mysql-connector==2.1.6
PyMySQL==0.9.2
Record==3.5
pipenv==2018.10.13
autopep8==1.4
pycodestyle==2.4.0
request==2018.11.20

